Remove specialisation requirement for stagger bar
It is not possible to fetch the player spec right away, and using
PLAYER_ENTERING_WORLD event also introduces some weitd behaviour, so
simply always creating the stagger element now in the case of a monk.

Later it would be nice to simply show/hide it based on spec + stance.